+++ FeedModules/Core Layer/FeedCoreModuleTests/FeedAPI/LoadFeedFromRemoteUseCaseTests.swift
@@ -157,33 +157,4 @@ class LoadFeedFromRemoteUseCaseTests: XCTestCase {
         return try! JSONSerialization.data(withJSONObject: json)
     }
     
-    // MARK: - Test SPY
-    private class HTTPClientSpy: HTTPClient {
-        
-        private struct Task: HTTPClientTask {
-            func cancel() {}
-        }
-        
-        // Store the request parameters in a tuple.
-        private var messages = [(url: URL, completion: (HTTPClient.Result) -> Void)]()
-        
-        var requestedURLs: [URL] {
-            return messages.map { $0.url }
-        }
-        
-        func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
-            messages.append((url, completion))
-            return Task()
-        }
-        
-        func complete(with error: Error, at index: Int = 0) {
-            messages[index].completion(.failure(error))
-        }
-        
-        func complete(withStatusCode code: Int, data: Data, at index: Int = 0) {
-            let response = HTTPURLResponse(url: requestedURLs[index], statusCode: code, httpVersion: nil, headerFields: nil)!
-            messages[index].completion(.success((data, response)))
-        }
-    }
-    
 }
